meme-generator/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
docx==0.2.4
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
pdftotext==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
